android-uxsdk-beta-test/src/androidTest/java/dji/ux/beta/cameracore/widget/cameraevsetting/CameraEVSettingWidgetModelTest.java
/*
 * Copyright (c) 2018-2021 DJI
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 *
 */

package dji.ux.beta.cameracore.widget.cameraevsetting;

import androidx.test.filters.SmallTest;
import androidx.test.ext.junit.runners.AndroidJUnit4;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.concurrent.TimeUnit;

import dji.common.camera.ExposureSettings;
import dji.common.error.DJIError;
import dji.keysdk.CameraKey;
import dji.thirdparty.io.reactivex.disposables.CompositeDisposable;
import dji.thirdparty.io.reactivex.observers.TestObserver;
import dji.thirdparty.io.reactivex.plugins.RxJavaPlugins;
import dji.thirdparty.io.reactivex.schedulers.TestScheduler;
import dji.thirdparty.io.reactivex.subscribers.TestSubscriber;
import dji.ux.beta.WidgetTestUtil;
import dji.ux.beta.core.base.DJISDKModel;
import dji.ux.beta.core.base.SchedulerProvider;
import dji.ux.beta.core.base.TestSchedulerProvider;
import dji.ux.beta.core.base.UXSDKError;
import dji.ux.beta.core.communication.ObservableInMemoryKeyedStore;

import static dji.common.camera.SettingsDefinitions.Aperture;
import static dji.common.camera.SettingsDefinitions.ExposureCompensation;
import static dji.common.camera.SettingsDefinitions.ExposureMode;
import static dji.common.camera.SettingsDefinitions.ExposureSensitivityMode;
import static dji.common.camera.SettingsDefinitions.ShutterSpeed;

/**
 * Test:
 * This class tests the public methods in {@link CameraEVSettingWidgetModel}
 * <p>
 * 1. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_isEditable_isUpdatedByManualMode()}
 * Test that the editability is updated when the camera is in manual mode.
 * 2. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_isEditable_isUpdatedByFixedEV()}
 * Test that the editability is updated when the camera has a fixed EV.
 * 3. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_getEVRange_isUpdated()}
 * Test that the EV range is updated.
 * 4. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_getCurrentEVPosition_isUpdated()}
 * Test that the current EV position is updated.
 * 5. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_getCurrentEVPosition_isUpdatedByManualMode()}
 * Test that the current EV position is updated when the camera is in manual mode.
 * 6. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_getCurrentEVPosition_isUpdatedByFixedEV()}
 * Test that the current EV position is updated when the camera has a fixed EV.
 * 7. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_eiMode_isUpdated()}
 * Test that the EI mode is updated.
 * 8. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_incrementEV_success()}
 * Test if the EV was incremented successfully.
 * 9. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_incrementEV_error()}
 * Test if incrementing the EV fails.
 * 10. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_incrementEV_outOfRange()}
 * Test if incrementing the EV fails due to being out of range.
 * 11. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_decrementEV_success()}
 * Test if the EV was decremented successfully.
 * 12. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_decrementEV_error()}
 * Test if decrementing the EV fails.
 * 13. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_decrementEV_outOfRange()}
 * Test if decrementing the EV fails due to being out of range.
 * 14. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_restoreEV_success()}
 * Test if the EV was restored successfully.
 * 15. {@link CameraEVSettingWidgetModelTest#cameraEVSettingWidgetModel_restoreEV_error()}
 * Test if restoring the EV fails.
 */
@RunWith(AndroidJUnit4.class)
@SmallTest
public class CameraEVSettingWidgetModelTest {
    private CompositeDisposable compositeDisposable;
    @Mock
    private DJISDKModel djiSdkModel;
    private CameraEVSettingWidgetModel widgetModel;
    private TestScheduler testScheduler;
    @Mock
    private ObservableInMemoryKeyedStore keyedStore;

    private ExposureCompensation[] exposureCompensationRange = new ExposureCompensation[]{
            ExposureCompensation.N_0_3,
            ExposureCompensation.N_0_0,
            ExposureCompensation.P_0_3};

    @Before
    public void beforeTest() {
        MockitoAnnotations.initMocks(this);
        RxJavaPlugins.reset();
        compositeDisposable = new CompositeDisposable();
        TestSchedulerProvider testSchedulerProvider = new TestSchedulerProvider();
        testScheduler = testSchedulerProvider.getTestScheduler();
        SchedulerProvider.setScheduler(testSchedulerProvider);
        widgetModel = new CameraEVSettingWidgetModel(djiSdkModel, keyedStore);
        WidgetTestUtil.initialize(djiSdkModel, widgetModel, true);

    }

    @Test
    public void cameraEVSettingWidgetModel_isEditable_isUpdatedByManualMode() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_MODE, 0),
                ExposureMode.MANUAL, 10, TimeUnit.SECONDS);
        widgetModel.setup();
        TestSubscriber<Boolean> testSubscriber =
                widgetModel.isEditable().test();

        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(0, isEditable -> isEditable);

        testScheduler.advanceTimeBy(10, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(1, isEditable -> !isEditable);

        widgetModel.cleanup();
        compositeDisposable.add(testSubscriber);
    }

    @Test
    public void cameraEVSettingWidgetModel_isEditable_isUpdatedByFixedEV() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.FIXED, 10, TimeUnit.SECONDS);
        widgetModel.setup();
        TestSubscriber<Boolean> testSubscriber =
                widgetModel.isEditable().test();

        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(0, isEditable -> isEditable);

        testScheduler.advanceTimeBy(10, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(1, isEditable -> !isEditable);

        widgetModel.cleanup();
        compositeDisposable.add(testSubscriber);
    }

    @Test
    public void cameraEVSettingWidgetModel_getEVRange_isUpdated() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 20, TimeUnit.SECONDS);
        widgetModel.setup();
        TestSubscriber<ExposureCompensation[]> testSubscriber =
                widgetModel.getEVRange().test();

        testScheduler.advanceTimeBy(25, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(1, result -> exposureCompensationRange == result);

        widgetModel.cleanup();
        compositeDisposable.add(testSubscriber);
    }

    @Test
    public void cameraEVSettingWidgetModel_getCurrentEVPosition_isUpdated() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 10, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.P_0_3, 20, TimeUnit.SECONDS);
        widgetModel.setup();
        TestSubscriber<Integer> testSubscriber =
                widgetModel.getCurrentEVPosition().test();

        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(0, result -> result == 15);

        testScheduler.advanceTimeBy(25, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(1, result -> result == 2);

        widgetModel.cleanup();
        compositeDisposable.add(testSubscriber);
    }

    @Test
    public void cameraEVSettingWidgetModel_getCurrentEVPosition_isUpdatedByManualMode() {
        ExposureSettings updatedExposureSettings = new ExposureSettings(Aperture.UNKNOWN,
                ShutterSpeed.UNKNOWN,
                0,
                ExposureCompensation.N_0_3);
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_MODE, 0),
                ExposureMode.MANUAL, 10, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 15, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_SETTINGS, 0),
                updatedExposureSettings, 20, TimeUnit.SECONDS);
        widgetModel.setup();
        TestSubscriber<Integer> testSubscriber =
                widgetModel.getCurrentEVPosition().test();

        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(0, result -> result == 15);

        testScheduler.advanceTimeBy(25, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(1, result -> result == 0);

        widgetModel.cleanup();
        compositeDisposable.add(testSubscriber);
    }

    @Test
    public void cameraEVSettingWidgetModel_getCurrentEVPosition_isUpdatedByFixedEV() {
        ExposureSettings updatedExposureSettings = new ExposureSettings(Aperture.UNKNOWN,
                ShutterSpeed.UNKNOWN,
                0,
                ExposureCompensation.FIXED);
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 10, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.FIXED, 20, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_SETTINGS, 0),
                updatedExposureSettings, 20, TimeUnit.SECONDS);
        widgetModel.setup();
        TestSubscriber<Integer> testSubscriber =
                widgetModel.getCurrentEVPosition().test();

        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(0, result -> result == 15);

        testScheduler.advanceTimeBy(25, TimeUnit.SECONDS);
        testSubscriber.assertValueAt(1, result -> result == 1);

        widgetModel.cleanup();
        compositeDisposable.add(testSubscriber);
    }

    @Test
    public void cameraEVSettingWidgetModel_eiMode_isUpdated() {
        // Use util method to set emitted value after given delay for given key
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_SENSITIVITY_MODE, 0),
                ExposureSensitivityMode.EI, 10, TimeUnit.SECONDS);
        // Setup the widget model after emitted values have been initialized
        widgetModel.setup();

        // Initialize a test subscriber that subscribes to the EI mode flowable from the model
        TestSubscriber<Boolean> testSubscriber = widgetModel.isEIMode().test();

        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        testSubscriber.assertValue(false);

        testScheduler.advanceTimeBy(15, TimeUnit.SECONDS);
        testSubscriber.assertValues(false, true);

        widgetModel.cleanup();
        compositeDisposable.add(testSubscriber);
    }

    @Test
    public void cameraEVSettingWidgetModel_incrementEV_success() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 5, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_3, 10, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedSetValue(djiSdkModel, CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_0, null);
        widgetModel.setup();

        testScheduler.advanceTimeBy(15, TimeUnit.SECONDS);
        TestObserver observer = widgetModel.incrementEV().test();
        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        observer.assertComplete();

        widgetModel.cleanup();
        compositeDisposable.add(observer);
    }

    @Test
    public void cameraEVSettingWidgetModel_incrementEV_error() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 5, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_3, 10, TimeUnit.SECONDS);
        UXSDKError uxsdkError = new UXSDKError(DJIError.COMMON_EXECUTION_FAILED);
        WidgetTestUtil.setEmittedSetValue(djiSdkModel, CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_0, uxsdkError);
        widgetModel.setup();

        testScheduler.advanceTimeBy(15, TimeUnit.SECONDS);
        TestObserver observer = widgetModel.incrementEV().test();
        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        observer.assertError(uxsdkError);
        widgetModel.cleanup();
        compositeDisposable.add(observer);
    }

    @Test
    public void cameraEVSettingWidgetModel_incrementEV_outOfRange() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 5, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.P_0_3, 10, TimeUnit.SECONDS);
        widgetModel.setup();

        testScheduler.advanceTimeBy(15, TimeUnit.SECONDS);
        TestObserver observer = widgetModel.incrementEV().test();
        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        observer.assertError(error -> ((Throwable) error).getMessage().equals("Exposure compensation is out of range"));
        widgetModel.cleanup();
        compositeDisposable.add(observer);
    }

    @Test
    public void cameraEVSettingWidgetModel_decrementEV_success() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 5, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_0, 10, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedSetValue(djiSdkModel, CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_3, null);
        widgetModel.setup();

        testScheduler.advanceTimeBy(15, TimeUnit.SECONDS);
        TestObserver observer = widgetModel.decrementEV().test();
        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        observer.assertComplete();

        widgetModel.cleanup();
        compositeDisposable.add(observer);
    }

    @Test
    public void cameraEVSettingWidgetModel_decrementEV_error() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 5, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_0, 10, TimeUnit.SECONDS);
        UXSDKError uxsdkError = new UXSDKError(DJIError.COMMON_EXECUTION_FAILED);
        WidgetTestUtil.setEmittedSetValue(djiSdkModel, CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_3, uxsdkError);
        widgetModel.setup();

        testScheduler.advanceTimeBy(15, TimeUnit.SECONDS);
        TestObserver observer = widgetModel.decrementEV().test();
        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        observer.assertError(uxsdkError);
        widgetModel.cleanup();
        compositeDisposable.add(observer);
    }

    @Test
    public void cameraEVSettingWidgetModel_decrementEV_outOfRange() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 5, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_3, 10, TimeUnit.SECONDS);
        widgetModel.setup();

        testScheduler.advanceTimeBy(15, TimeUnit.SECONDS);
        TestObserver observer = widgetModel.decrementEV().test();
        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        observer.assertError(error -> ((Throwable) error).getMessage().equals("Exposure compensation is out of range"));
        widgetModel.cleanup();
        compositeDisposable.add(observer);
    }

    @Test
    public void cameraEVSettingWidgetModel_restoreEV_success() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 5, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_3, 10, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedSetValue(djiSdkModel, CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_0, null);
        widgetModel.setup();

        testScheduler.advanceTimeBy(15, TimeUnit.SECONDS);
        TestObserver observer = widgetModel.restoreEV().test();
        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        observer.assertComplete();

        widgetModel.cleanup();
        compositeDisposable.add(observer);
    }

    @Test
    public void cameraEVSettingWidgetModel_restoreEV_error() {
        initKeys();
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0),
                exposureCompensationRange, 5, TimeUnit.SECONDS);
        WidgetTestUtil.setEmittedValue(widgetModel, djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_3, 10, TimeUnit.SECONDS);
        UXSDKError uxsdkError = new UXSDKError(DJIError.COMMON_EXECUTION_FAILED);
        WidgetTestUtil.setEmittedSetValue(djiSdkModel, CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0),
                ExposureCompensation.N_0_0, uxsdkError);
        widgetModel.setup();

        testScheduler.advanceTimeBy(15, TimeUnit.SECONDS);
        TestObserver observer = widgetModel.restoreEV().test();
        testScheduler.advanceTimeBy(5, TimeUnit.SECONDS);
        observer.assertError(uxsdkError);
        widgetModel.cleanup();
        compositeDisposable.add(observer);
    }

    @After
    public void afterTest() {
        RxJavaPlugins.reset();
        compositeDisposable.dispose();
    }

    private void initKeys() {
        WidgetTestUtil.setEmptyValue(widgetModel,
                djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION, 0));
        WidgetTestUtil.setEmptyValue(widgetModel,
                djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_MODE, 0));
        WidgetTestUtil.setEmptyValue(widgetModel,
                djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_COMPENSATION_RANGE, 0));
        WidgetTestUtil.setEmptyValue(widgetModel,
                djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_SETTINGS, 0));
        WidgetTestUtil.setEmptyValue(widgetModel,
                djiSdkModel,
                CameraKey.create(CameraKey.EXPOSURE_SENSITIVITY_MODE, 0));
        WidgetTestUtil.setEmptyLensValues(widgetModel, djiSdkModel, 0);
    }
}
